Next
a crucible is heated red hot and the dry powder mixture is slowly added spoonful by spoonful with a cover placed on top with each new addition. A scoria will form at the top and regulus will fall to the bottom of the crucible. Pour the melt into a hot cone-shaped iron mold and tap the sides to help the regulus fall to the bottom. Once cool, the regulus can be separated from the scoria easily by a hammer blow. Save the scoria aside; it contains the “seed of gold”.