Repeat this purification two more times at least
The next step is to form the Martial Regulus. The recommended proportions are 1 part regulus to 1/10th part iron or steel nails. Iron is said to have a volatile magnetic quality and steel, a fixed magnetic quality. Steel is preferred here, but not stainless steel. Heat the nails red hot in a crucible, then slowly add the regulus as a powder. Keep in flux for 15 minutes, then pour into the iron cone which has been preheated. When cool, break the regulus away from the remaining undissolved nails. Purify the regulus as described above until you see the starry crystal surface.