Ideally
we should use a very pure Native Cinnabar for this process, but if that is not available, we can use high purity metallic mercury (triple distilled) from the chemical supply store. It is a good idea to purify it in the old ways prior to use. It’s not so much aimed at getting the mercury any more pure as at exalting it alchemically, opening the body to receive new life.