Now
cover the mercury with dried and powdered sea salt. Mix these two completely using a mortar and pestle. Depending on the impurities in the mercury, the salt will darken even to blackness. Wash this dirty salt out using rainwater and repeat the salt washing once or twice. The mercury will retain some of the subtle essence of the salt which is important for the rest of the processes.