Put the mercury into a mortar
and add an equal amount of sea salt. Saturate this mixture with strong distilled vinegar (about 10% acetic acid) and begin vigorously mixing the mass for about 10 minutes. Now wash this salt out with rainwater until the mercury appears shiny and bright. Press it through a chamois; there should be no residue left in the chamois. The mercury should leave no trail when rolled across a smooth surface and should be without any appearance of scum at the surface.