Mix the mercury
with an equal amount of native sulfur in a mortar, grinding it together well. The matter will turn black during this process, forming mercury sulfide. Examine the matter closely with a magnifier. There should be no tiny globules of mercury remaining. Grind with additional sulfur if there are. The resulting black mass is a crude form of cinnabar, which is the sulfide ore of mercury.