Begin the distillation
slowly at first, and then gradually raise the heat until metallic mercury distills over. The temperature will be about 320 to 380 deg C as the mercury distills. When the distillation is complete, remove the receiver first, before removing the heat, or the cooling retort may draw the water into itself and explode.