The silver
then often called the “Doves of Diana”, acts as a medium to transfer the life force of the antimony into the mercury. After the distillation, the residue of silver which remains in the retort, called now the “Dead Doves of Diana”, is cleaned and used again each time with the addition of fresh regulus.