The purification
of the precious metals begins with the preparation of sea salt. The salt is fused in a crucible at about 800 deg C, then dissolved in rainwater. Filter the solution into a distillation train and distill out 2/3 of the water. Save this distilled water aside, and pour the remaining salt concentrate into a dish.