For the White Work, silver metal is dissolved
in Aqua Fortis or nitric acid. Filter the solution through glass wool and dilute it with about ten times as much water. Now pour the solution into a clean dish made of copper. Some of the copper will dissolve and fine crystals of silver metal will form on the bottom of the dish. Wash this fine silver powder several times with rainwater, then dry it.