Take the ball of amalgam remaining
in the chamois and grind it together with the previously purified salt. Add the salt in small amounts and continue grinding until you cannot see the metallic amalgam in
the powder. Place this powder into a distillation train and distill out the mercury. The residue remaining in the distillation flask is rinsed several times to dissolve and remove the salt. A very fine “silver calx” will remain.