Now that we have prepared the metal calx and the Animated Mercury,
we are ready to proceed with the next phase of the work. This part of the work consists of planting our precious metal “seed” into the Sophic Mercury. Take one part of silver calx (for the White Work) or one part of gold calx (for the Red Work) and grind it together with two parts of the Sophic Mercury for about ten hours. The resulting amalgam is then washed with rainwater until no more blackness comes out in the water and the amalgam appears clean and bright.