Place 3 parts of the Sophic Gold into a clean digestion flask
as above and heat it until you can barely stand to touch it for very long. Using a heated glass funnel, you must add 1 part of preheated Sophic mercury in small portions. Add about the quantity of a pea each time and allow about half an hour to pass before the next addition. Each time the moisture should be allowed to vapor away.