Now take 1 part of the Sophic Gold reduced to powder and place it into a clean digestion flask.
Add two parts of Sophic Mercury all at once, then gently warm the flask to exhale all of the moisture. Seal the flask and continue the digestion until the powder and mercury form a fluid amalgam. Pour the amalgam into a clean glass mortar and begin gently grinding it with rainwater until no more blackness enters the water. When the wash water comes off clear, pour the amalgam into a dish and let it dry in the sun.