Digest with a gentle heat in a sandbath
for 30 days and the darkness will begin to lighten. Continue the digestion for another 30 days and the matter will take on a greenish hue, then a blue green, similar to a
peacock’s tail, which is what this phase of the work is called. As the digestion continues, the matter will become increasingly yellow, then reddish. At last, the entire mass will become a beautiful red powder which is called the “Red Sulfur of the Philosophers”.