The process of transmutation of metals into gold is called Projection.
Making a judgment of how much metal your stone will transmute requires a little experimentation. Take about 4 parts of pure gold and melt it in a clean crucible. Cast in 1 part of your Red Stone wrapped in beeswax and keep the matter in flux for half an hour. When cool, break out the resulting red glassy mass from the crucible and powder it.