Now heat 100 parts common mercury
in a crucible and cast in 1 part of the Red Stone wrapped in beeswax. Keep the matter in flux for half an hour, then cast it out into an ingot mold. If the metal is still brittle, you must re-melt it with more mercury. If all of the metal did not change into gold, you must add more of the Stone. After a few trials, you will be able to judge the proportions in future transmutations.