In order to multiply the Stone in quantity and virtue,
melt 1 part of it with 2 parts pure gold. Powder the resulting glassy mass and place it into a digestion flask. Add Sophic mercury until it becomes a soft mass, then seal the flask. Digest at a gentle heat and it will become fixed into the Red Stone in a short time.