White Stone
The final phase of this process is the preparation of the White and Red Stones from this White Sulfur. For the White Stone we use pure silver metal and for the Red Stone we use Pure Gold. In each case the method is the same, so we will only describe here the preparation of the Red Stone. Place the White Sulfur into a digestion flask and add small portions of Sophic Mercury until the matter is a soft dark mass. Warm the flask to exhale all moisture, then seal it tightly.